Fix file naming code smells - #664
