+++ detekt-api/src/test/kotlin/io/gitlab/arturbosch/detekt/api/Compiler.kt
