+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/DetektResult.kt
