+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/ModificationNotification.kt
