+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/Case.kt
